+++ amaptor/classes/layout.py
@@ -3,6 +3,8 @@ log = logging.getLogger("amaptor")
 
 from amaptor.classes.map_frame import MapFrame
 from amaptor.version_check import PRO
+from amaptor.errors import MapFrameNotFoundError
+
 class Layout(object):
 	"""
 		Replicates Layouts so that we can do some nice things behind the scenes.
@@ -15,6 +17,7 @@ class Layout(object):
 		self._layout_object = layout_object
 		self.project = project
 		self.frames = [MapFrame(frame, self) for frame in self._layout_object.listElements("MAPFRAME_ELEMENT")]  # frames connect back to maps, this connects to a project
+		self.elements = self.list_elements()
 
 	@property
 	def name(self):
@@ -34,6 +37,9 @@ class Layout(object):
 		else:
 			self.project.primary_document.title = value
 
+	def list_elements(self):
+		return self._layout_object.listElements()
+
 	def export_to_png(self, out_path, resolution=300):
 		"""
 			Currently Pro only - needs refactoring to support ArcMap and Pro (should export map document in ArcMap).
@@ -43,6 +49,50 @@ class Layout(object):
 		:return:
 		"""
 		self._layout_object.exportToPNG(out_path, resolution)
+
+	def find_map_frame(self, name):
+		"""
+			Finds the map frame with a given name
+		:param name: the name of the frame to find
+		:return: MapFrame object
+		"""
+
+		if PRO:
+			for frame in self.frames:
+				if name == frame.name:
+					return frame
+			else:
+				raise MapFrameNotFoundError(name=name)
+
+	def find_element(self, name):
+		"""
+			Finds the first element matching the provided name
+		:param name:
+		:return:
+		"""
+		for element in self.elements:
+			if element.name == name:
+				return name
+
+	def toggle_element(self, name, visibility="TOGGLE"):
+		"""
+			Given an element name, toggles, makes visible, or makes invisible that element.
+		:param name:
+		:param visibility: Controls the action. Valid values are boolean (True, False) or the keyword "TOGGLE" which
+			switches its current visibility state.
+		:return:
+		"""
+
+		element = self.find_element(name)
+		if visibility is True or visibility is False:
+			element.visible = visibility
+		elif visibility == "TOGGLE":
+			if element.visible is True:
+				element.visible = False
+			else:
+				element.visible = True
+		else:
+			raise ValueError("parameter visibility must be either a boolean value, (True, False) or the keyword \"TOGGLE\".")
 
 	def export_to_pdf(self, out_path, **kwargs):
 		self._layout_object.exportToPDF(out_path, **kwargs)